Consolidate utils tests into elife-utils spec & html fixtures files
