--- test/utils.html
@@ -1,87 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-  <title>Mocha Test Runner</title>
-
-  <link rel="stylesheet" href="../node_modules/mocha/mocha.css">
-</head>
-<body>
-
-    <!--
-      Test output
-    -->
-    <div id="mocha"></div>
-
-    <!--
-      Fixtures
-    -->
-
-    <!--isIdOfOrWithinSection-->
-    <div class="isIdOfOrWithinSection">
-      <div class="target" id="isIdOfOrWithinSection_targetHit">
-        <div class="target-child" id="isIdOfOrWithinSection_targetHit_child">
-          <div class="target-grand-child"  id="isIdOfOrWithinSection_targetHit_grandChild">
-          </div>
-        </div>
-      </div>
-      <div class="target-miss" id="isIdOfOrWithinSection_targetMiss"></div>
-    </div>
-
-    <!--areElementsNested-->
-    <div class="areElementsNested">
-      <div class="target">
-        <div class="target-child">
-          <div class="target-grand-child">
-          </div>
-        </div>
-      </div>
-      <div class="target-miss"></div>
-    </div>
-
-    <!--getOrdinalAmongstSiblingElements-->
-    <div class="getOrdinalAmongstSiblingElements">
-      <div class="parent">
-        <div class="child-1a"></div>
-        <div class="child-2"></div>
-        <div class="child-3"><span class="child-1b"></span></div>
-      </div>
-    </div>
-
-    <!--buildElement-->
-    <div class="buildElement">
-      <div class="parent first-parent">
-        <div class="pre-existing-sibling-1"></div>
-        <div class="pre-existing-sibling-2"></div>
-      </div>
-      <div class="parent this-second-parent-should-not-match not-a-child-of-first-parent">
-    </div>
-
-
-    <!--
-      Load test code
-    -->
-    <script src="../node_modules/mocha/mocha.js"></script>
-    <script src="../node_modules/chai/chai.js"></script>
-    <script src="../node_modules/sinon/pkg/sinon.js"></script>
-
-    <!-- Initialise test suite -->
-    <script>mocha.setup('bdd')</script>
-
-    <!--
-      Test spec
-    -->
-    <script src="build/utils.spec.js"></script>
-
-    <!--
-      Begin!
-    -->
-    <script>
-      mocha.run();
-    </script>
-
-</body>
-</html>
